refactor: grouped components by folders
